Feature: Window server
Contraption is a window server that performs fully server-side
rendering. Once windows and widgets are handed over to the server, it
will perform rendering and scene handling completely on its own.

As of now there is no way how window server can notify applications
back. This is TBD. What works is window and widget creation and basic
window management interaction.

User can:
* raise windows by clicking them
* move windows by dragging their title bars
* open menus by clicking menu bar items

Screen update is currently working only partially as there is rework of
drawing from back-to-front to front-to-back.

Only a limited subset of widgets is supported right now:

* plain text
* buttons
* menu items
* title bars

Their behavior and visual is fixed by the implementation of the window
server.