Contraption: Cleanup, fix menu rendering
Split code of contraption into more files. Isolate window-related code
into own file and rendering primitives code into another file. This is a
preparation for damage-driven front-to-back rendering.

Add "event callback" to gadgets. This will be used for all kinds of
internal events dispatched to gadgets. This is different than events
dispatched to the clients, which are defined by the client.

Fix stray menu bug that caused menu to remain displayed after it was
activated even if the window which owned it lost focus.

Fix stray active menu item that caused menu items to remain activated if
menu it opened was closed by clicking outside of menu.